Moved Round and Game generation logic into it's own class which is registered, injected and used by RoundManagement controller.
