+++ LZRNS.Core/IRoundGenerator.cs
@@ -1,0 +1,11 @@
+﻿using LZRNS.DomainModel.Models;
+using LZRNS.DomainModels.Models;
+using System.Collections.Generic;
+
+namespace LZRNS.Core
+{
+    public interface IRoundGenerator
+    {
+        IEnumerable<Round> GenerateRoundsWithGames(List<Team> teams, LeagueSeason leagueSeason);
+    }
+}
