+++ LZRNS.Core/LZRNS.Core.csproj
@@ -94,7 +94,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Handlers\SearchHandler.cs" />
+    <Compile Include="IRoundGenerator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RoundGenerator.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -103,6 +105,10 @@
     <ProjectReference Include="..\LZRNS.Common\LZRNS.Common.csproj">
       <Project>{44ea5768-a92d-4380-848e-a4b2c07158d6}</Project>
       <Name>LZRNS.Common</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\LZRNS.DomainModels\LZRNS.DomainModels.csproj">
+      <Project>{584350e5-f1db-4334-b57a-7e1fa59a6a25}</Project>
+      <Name>LZRNS.DomainModels</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
