+++ LZRNS.Web/LZRNS.Web.csproj
@@ -761,6 +761,10 @@
       <Project>{44ea5768-a92d-4380-848e-a4b2c07158d6}</Project>
       <Name>LZRNS.Common</Name>
     </ProjectReference>
+    <ProjectReference Include="..\LZRNS.Core\LZRNS.Core.csproj">
+      <Project>{e8c6667b-38c8-4258-9313-f8b83d79e43a}</Project>
+      <Name>LZRNS.Core</Name>
+    </ProjectReference>
     <ProjectReference Include="..\LZRNS.DomainModels\LZRNS.DomainModels.csproj">
       <Project>{584350e5-f1db-4334-b57a-7e1fa59a6a25}</Project>
       <Name>LZRNS.DomainModels</Name>
